wheel_speed_analyser/requirements.txt
decorator==4.0.10
ipython==5.1.0
ipython-genutils==0.1.0
numpy==1.11.2
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.8
ptyprocess==0.5.1
Pygments==2.1.3
pyserial==3.2.1
python-osc==1.6.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
uflash==1.0.7
wcwidth==0.1.7
